renamed resolvers.js -> resolversFactory.js
